apps/backend/requirements.txt
Flask==3.1.1
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Pillow==11.0.0
python-magic==0.4.27
Werkzeug==3.1.3
PyJWT==2.10.1
pytest==8.3.5
pytest-flask==1.3.0
pytest-mock==3.14.1 